# docs/swagger.yaml
definitions:
  dto.BarcodeListRequest:
    properties:
      barcodes:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - barcodes
    type: object
  dto.BarcodeOperationResponse:
    properties:
      barcodes:
        items:
          type: string
        type: array
      processed_count:
        type: integer
    type: object
...
  dto.CreateFleetRequest:
    properties:
      driver_name:
        type: string
      max_weight_kg:
        type: number
      plate_number:
        type: string
    required:
    - max_weight_kg
    - plate_number
    type: object
...
  dto.FillingBatchDetailResponse:
    properties:
      barcode_sn:
        type: string
      cylinder_id:
        type: string
      id:
        type: string
    type: object
  dto.FillingBatchResponse:
    properties:
      batch_number:
        type: string
      created_at:
        type: string
      cylinder_qty:
        type: integer
      details:
        items:
          $ref: '#/definitions/dto.FillingBatchDetailResponse'
        type: array
      gas_type:
        type: string
      id:
        type: string
      item_id:
        type: string
      item_name:
        type: string
      notes:
        type: string
      status:
        type: string
    type: object
  dto.FleetResponse:
    properties:
      driver_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_weight_kg:
        type: number
      plate_number:
        type: string
    type: object
  dto.HydrotestDueCylinder:
    properties:
      barcode_sn:
        type: string
      expiry_date:
        type: string
      id:
        type: string
      is_expired:
        type: boolean
      last_hydrotest_date:
        type: string
      status:
        type: string
    type: object
  dto.HydrotestDueResponse:
    properties:
      due_within_days:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.HydrotestDueCylinder'
        type: array
    type: object
...
  dto.Message:
    properties:
      message:
        type: string
    type: object
...
  dto.PaginatedFillingBatchList:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.FillingBatchResponse'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.PaginatedFleetList:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.FleetResponse'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
...
  dto.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RecordHydrotestRequest:
    properties:
      last_hydrotest_date:
        type: string
      notes:
        type: string
    required:
    - last_hydrotest_date
    type: object
...
  dto.StockOpnameRequest:
    properties:
      actual_quantity:
        minimum: 0
        type: integer
      item_id:
        type: string
      notes:
        type: string
    required:
    - item_id
    type: object
  dto.StockOpnameResponse:
    properties:
      item_id:
        type: string
      item_name:
        type: string
      quantity_after:
        type: integer
      quantity_before:
        type: integer
      quantity_delta:
        type: integer
    type: object
  dto.SubmitFillingBatchRequest:
    properties:
      barcodes:
        items:
          type: string
        minItems: 1
        type: array
      item_id:
        type: string
      notes:
        type: string
    required:
    - barcodes
    - item_id
    type: object
...
  dto.UpdateFleetRequest:
    properties:
      driver_name:
        type: string
      is_active:
        type: boolean
      max_weight_kg:
        type: number
    required:
    - max_weight_kg
    type: object
...
  global.Response-dto_BarcodeOperationResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.BarcodeOperationResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_FillingBatchResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.FillingBatchResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  global.Response-dto_FleetResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.FleetResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  global.Response-dto_HydrotestDueResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.HydrotestDueResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_Message:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Message'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_PaginatedFillingBatchList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.PaginatedFillingBatchList'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  global.Response-dto_PaginatedFleetList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.PaginatedFleetList'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_StockOpnameResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.StockOpnameResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    email: debidarmawan1998@gmail.com
    name: DeboZero Corp Tech Team
  description: This is an API documentation of Progas WMS
  title: PROGAS WMS Backend API
  version: "1.0"
paths:
...
  /inbound/empty-receive:
    post:
      consumes:
      - application/json
      description: Receive empty cylinders returned from customer (OUTSTANDING/IN_TRANSIT
        → EMPTY)
      parameters:
      - description: Barcode list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BarcodeListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_BarcodeOperationResponse'
      security:
      - Bearer: []
      summary: Empty cylinder receiving
      tags:
      - Inbound
  /inventory/spareparts/stock-opname:
    post:
      description: Set actual spare part quantity (adjusts stock to counted value)
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StockOpnameRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_StockOpnameResponse'
      security:
      - Bearer: []
      summary: Spare part stock opname
      tags:
      - Inventory
...
  /logistics/fleet:
    get:
      consumes:
      - application/json
      description: List fleet vehicles with pagination and search
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Search by plate number or driver name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_PaginatedFleetList'
      security:
      - Bearer: []
      summary: List fleet vehicles
      tags:
      - Logistics
    post:
      consumes:
      - application/json
      description: Register a fleet vehicle with max weight capacity
      parameters:
      - description: Create fleet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFleetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Create fleet vehicle
      tags:
      - Logistics
  /logistics/fleet/{id}:
    get:
      consumes:
      - application/json
      description: Get fleet vehicle detail
      parameters:
      - description: Fleet vehicle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_FleetResponse'
      security:
      - Bearer: []
      summary: Get fleet vehicle by id
      tags:
      - Logistics
    put:
      consumes:
      - application/json
      description: Update fleet vehicle driver, max weight, or active status
      parameters:
      - description: Fleet vehicle ID
        in: path
        name: id
        required: true
        type: string
      - description: Update fleet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFleetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Update fleet vehicle
      tags:
      - Logistics
...
  /maintenance/cylinders/{id}/hydrotest:
    post:
      description: Update last hydrotest date and set status to MAINTENANCE
      parameters:
      - description: Cylinder ID
        in: path
        name: id
        required: true
        type: string
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordHydrotestRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Record cylinder hydrotest
      tags:
      - Maintenance
  /maintenance/hydrotest/due:
    get:
      parameters:
      - description: Due within days (default 30)
        in: query
        name: days
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_HydrotestDueResponse'
      security:
      - Bearer: []
      summary: List cylinders due for hydrotest
      tags:
      - Maintenance
...
  /production/filling-batches:
    get:
      consumes:
      - application/json
      description: List filling batches with pagination and search
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Search by batch number or gas type
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_PaginatedFillingBatchList'
      security:
      - Bearer: []
      summary: List filling batches
      tags:
      - Production
    post:
      consumes:
      - application/json
      description: Create and complete a filling batch atomically (validates status
        & cross-gas, sets cylinders to FILLED; use post-fill QC for READY)
      parameters:
      - description: Filling batch request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitFillingBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_FillingBatchResponse'
      security:
      - Bearer: []
      summary: Submit filling batch
      tags:
      - Production
  /production/filling-batches/{id}:
    get:
      consumes:
      - application/json
      description: Get filling batch detail including scanned cylinders
      parameters:
      - description: Filling batch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_FillingBatchResponse'
      security:
      - Bearer: []
      summary: Get filling batch by id
      tags:
      - Production
  /production/qc/post-fill:
    post:
      consumes:
      - application/json
      description: Pass post-fill inspection after gas filling (FILLED → READY)
      parameters:
      - description: Barcode list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BarcodeListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_BarcodeOperationResponse'
      security:
      - Bearer: []
      summary: Post-fill QC
      tags:
      - Production
  /production/qc/pre-fill:
    post:
      consumes:
      - application/json
      description: Pass pre-fill inspection (EMPTY → READY_TO_FILL)
      parameters:
      - description: Barcode list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BarcodeListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_BarcodeOperationResponse'
      security:
      - Bearer: []
      summary: Pre-fill QC
      tags:
      - Production
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"